Fix tests and bump version
