+++ CONTRIBUTING.md
@@ -4,8 +4,8 @@ We welcome contributions and here we lay out some guidelines which should be fol
 
 ## Contribution process
 
-Commits should not be pushed directly to the master branch but should instead be merged from feature branches following via pull/merge requests. To track 
-tasks, features, bugs and enhancements should have a corresponding issue which explains the motivation and logic used in the committed code/documentation.
+Commits should not be pushed directly to the master branch but should instead be merged from feature branches following via pull/merge requests.
+To track tasks, features, bugs and enhancements should have a corresponding issue which explains the motivation and logic used in the committed code/documentation.
 The steps in the full process from creating an issue to merging are:
 
 1. Create issue with quick description of feature, bug, enhancement etc.. This will be assigned an issue number
