+++ Project.toml
@@ -1,7 +1,7 @@
 name = "QXTools"
 uuid = "84f0eee1-10ae-40da-994b-b9d0e53829ae"
 authors = ["QuantEx team"]
-version = "0.1.9"
+version = "0.2.0"
 
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
@@ -12,12 +12,12 @@ JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 LightGraphs = "093fc24a-ae57-5d10-9952-331d41423f4d"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
-MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 QXContexts = "04c26001-d4a1-49d2-b090-1d469cf06784"
 QXGraphDecompositions = "b15f1ded-d19a-4aca-a940-1991fcfc7750"
 QXTns = "995e1dad-72e3-4b97-b122-7deaeb8b44f9"
 QXZoo = "f27fdc93-0c88-4b5a-91cb-e8275adac0f6"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 TensorOperations = "6aa20fa7-93e2-5fca-9bc0-fbd0db3c71a2"
 TestSetExtensions = "98d24dd4-01ad-11ea-1b02-c9a08f80db04"
 YAML = "ddb6d928-2868-570f-bddf-ab3f9cf99eb6"
@@ -28,12 +28,14 @@ ArgParse = "1.1"
 DataStructures = "0.18"
 FileIO = "1.6"
 JLD2 = "0.4"
-LightGraphs = "1.3.5"
+LightGraphs = "1.3"
+QXContexts = "0.1"
 QXGraphDecompositions = "0.1"
 QXTns = "0.1"
 QXZoo = "0.1"
+Reexport = "1.1"
 TensorOperations = "3.1"
-TestSetExtensions = "2.0.0"
+TestSetExtensions = "2.0"
 YAML = "0.4"
 julia = "1.6"
