+++ src/QXTools.jl
@@ -1,5 +1,6 @@
 module QXTools
 using Logging
+using Reexport
 import QXZoo
 import DataStructures: OrderedDict
 
@@ -10,8 +11,12 @@ include("tn_conversion.jl")
 # contraction planning
 include("contraction_planning.jl")
 # dsl and parameter files
-include("dsl/dsl.jl")
+include("compute_graph/compute_graph.jl")
+# functions for output arguments and parsing
+include("outputs.jl")
 # simulation utilities
 include("simulation.jl")
+
+@reexport using QXTools.Circuits
 
 end
